LayoutTests/fast/dom/Window/Location/location-override-toString-on-proto-using-defineGetter.html

<html>
<head>
    <link rel="stylesheet" href="../../../js/resources/js-test-style.css">
    <script src="../../../js/resources/js-test-pre.js"></script>
    <script src="resources/location-tests-functions.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
    window.location.__proto__.__defineGetter__("toString", function () {
        return function() { return "haxored"; }
    });

    var result = normalizeURL(String(window.location));
    var correctValue = normalizeURL(document.URL);
    shouldBe("result", "correctValue");

    successfullyParsed = true;
</script>
<script src="../../../js/resources/js-test-post.js"></script>
</body>
</html>
